modif header

+++ devoir_1.h
@@ -8,7 +8,7 @@
 
 
 void tridiagonalize(double *A, int n, int k, double *d, double *e);
-int step_qr_tridiag(double *d, double *e, double m, double eps);
+int step_qr_tridiag(double *d, double *e, int m, double eps);
 int qr_eigs_(double *A, int n, int k, double eps, int max_iter, double *d);
 
 #endif
